refactor: move select components to input directory
